Added try/catch to address crashing issue. This issue arose when the user preformed (no match,select feature,next feature, previous feature)...Resulting in a crash.
The try/catch will assess if a row is deletable from the dataframe, and if it is, then delete, if not, then the app will not crash, but will allow user to adjust select points/etc.

Breifly edited UI (shifted text field), and updated a column value which is entered into dataframe.