Updated shiny app to have proper fields which are included in the "stems" output from "main" script.
Also increased the search area of the matching window to -4m & +4m from the boundaries of the bbox of the focused feature. (Better "context")

Should update so that all potential LiDAR features are displayed, to provide better context during each matching step. Currently "blind" to later or prior LiDAR stems.